+++ CHANGES.md
@@ -4,6 +4,8 @@
 * fixed `LocalPool` usage in tests
 * fixed config load plugin call
 * added latest `ControlUtility` commands for AI 2.9.0
+* `download_from_host` now supports lists of remote files
+* added `download_from_nodes`
 
 #### *0.6.2* @ 2020-06-05
 * added license banners to all sources files